NewEvent persistance working

+++ Backend/src/main/java/com/unevento/api/records/NewEvent.java
@@ -2,5 +2,5 @@ package com.unevento.api.records;
 
 import java.util.Date;
 
-public record NewEvent(Long userID, String nombre, String descripcion, String lugar, String categoria, String Facultad, Date fechaEvento, int capacidad, String hora) {
+public record NewEvent(Long userID, String nombre, String descripcion, String lugar, String categoria, String Facultad, Date fechaEvento, Long capacidad, String hora) {
 }
